Added fork to OCR
